[REFACTOR] Move import_submodules() to `neurite` __init__.py` and have submodules import it.
